tech(cli): rename log to logger to separate from server logger
