+++ pkg/cmd/grafana-cli/logger/logger.go
@@ -1,4 +1,4 @@
-package log
+package logger
 
 import (
 	"fmt"
